Workaround for a concurrency issue for buggy java STL impls (e.g. GNU Classpath 0.91)
In the issue it seems like the thread that is performing a blocking read
operation (the AMQConnection.MainLoop class) is also hindering write
operations to the TCP socket (e.g. a basic publish on a channel).

The workaround is to make the blocking read operations time out faster. This
lets any pending write operations take place. Currently the time out interval
is hard coded to 250 ms (however this should be dependent on RTT, but also
should not be too big or otherwise write ops will stall).

However, in this previous implementation the timeout interval was hardcoded to
a fourth of the heartbeat interval. Which meant that after four timed out read
operations, a missed heartbeat would be registered. This has been decoupled in
this commit as well.

One other unresolved issue is the AMQP handshake interval, this workaround
completely ignores this handshake interval.